requirements.txt
aiohttp==3.7.4.post0
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
chart-studio==1.1.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docopt==0.6.2
docutils==0.16
entrypoints==0.3
fsspec==2021.5.0
gcsfs==2021.5.0
gitdb==4.0.5
GitPython==3.1.13
google-auth==1.30.0
google-auth-oauthlib==0.4.4
idna==2.10
install==1.3.4
ipykernel==5.5.0
ipython==7.20.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keyring==22.0.1
kiwisolver==1.3.1
lxml==4.6.3
MarkupSafe==1.1.1
matplot==0.1.9
matplotlib==3.3.4
mistune==0.8.4
multidict==5.1.0
multitasking==0.0.9
mysql-connector-python==8.0.21
nbclient==0.5.2
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==5.7.8
numpy==1.20.1
oauthlib==3.1.0
packaging==20.9
pandas==1.2.2
pandocfilters==1.4.3
parso==0.8.1
patsy==0.5.1
pickleshare==0.7.5
Pillow==8.1.0
pipreqs==0.4.10
pkginfo==1.7.0
plotly==4.14.3
prometheus-client==0.9.0
prompt-toolkit==3.0.16
protobuf==3.15.1
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydeck==0.6.1
Pygments==2.8.0
pyloco==0.0.139
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pywin32-ctypes==0.2.0
pywinpty==0.5.7
pyzmq==22.0.3
readme-renderer==28.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
retrying==1.3.3
rfc3986==1.4.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.1
seaborn==0.11.1
Send2Trash==1.5.0
SimpleWebSocketServer==0.1.1
six==1.15.0
sklearn==0.0
smmap==3.0.5
statsmodels==0.12.2
streamlit==0.82.0
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
toolz==0.11.1
tornado==5.1.1
tqdm==4.57.0
traitlets==5.0.5
twine==3.3.0
typing==3.7.4.3
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.3
ushlex==0.99.1
validators==0.18.2
watchdog==2.0.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
widgetsnbextension==3.5.1
yarg==0.1.9
yarl==1.6.3
yfinance==0.1.59
